pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "cosmic-mycelium"
version = "0.1.0"
description = "Cosmic Mycelium — A self-evolving, ternary-model-based silicon-based lifeform core"
readme = "README.md"
license = {text = "AGPL-3.0"}
requires-python = ">=3.12,<3.14"
authors = [
    {name = "Stardust & Xuanji", email = "cosmic.mycelium@unknown"},
]
keywords = ["ai", "silicon-life", "cognitive-core", "mycelium-network", "ternary-model"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Distributed Computing",
]

# Core dependencies — pinned to exact versions for reproducibility
dependencies = [
    # Async runtime
    "anyio==4.9.0",
    "trio==0.28.0",
    # Data validation & serialization
    "pydantic==2.11.4",
    "pydantic-settings==2.9.1",
    "msgspec==0.19.0",
    # Numerical & scientific computing
    "numpy==2.2.6",
    "scipy==1.15.3",
    # Vector operations & similarity
    "hnswlib==0.8.0",
    "faiss-cpu==1.11.0",
    # Machine learning (lightweight)
    "torch==2.7.0",
    # Liquid Neural Networks (LTC) — Phase 4.2 cross-civilization semantic mapping
    "ncps>=1.0.0",
    "tensorflow>=2.16.0",
    # Quantum-inspired ternary computing (1.6.0+ supports Python 3.13)
    "qiskit>=1.6.0",
    # Cryptography & security
    "cryptography==44.0.3",
    "pyjwt==2.10.1",
    # Networking & communication
    "redis==6.1.0",
    "kafka-python==2.0.2",
    "aio-pika==9.5.5",
    # Configuration management
    "pyyaml==6.0.2",
    "tomli==2.2.1",
    "tomli-w==1.2.0",
    # Monitoring & observability
    "prometheus-client==0.22.0",
    "opentelemetry-api==1.33.0",
    "opentelemetry-sdk==1.33.0",
    # HTTP server for metrics/health
    "aiohttp==3.11.18",
    # Logging
    "structlog==25.3.0",
    "rich==14.0.0",
    # Testing
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "pytest-cov==6.1.1",
    "pytest-xdist==3.6.1",
    "hypothesis==6.152.1",
    # Type checking
    "mypy==1.16.0",
    # Code quality
    "ruff==0.11.10",
    "isort==6.0.1",
    "black==25.1.0",
    "codespell==2.4.1",
    # Utilities
    "click==8.2.0",
    "python-dotenv==1.1.0",
    "tenacity==9.1.2",
]

[project.optional-dependencies]
dev = [
    "pre-commit==4.2.0",
    "pytest-benchmark==5.1.0",
    "pytest-mock==3.14.0",
    "factory-boy==3.3.3",
    "faker==37.3.0",
]
docs = [
    "sphinx==8.2.3",
    "sphinx-rtd-theme==3.0.2",
    "myst-parser==4.0.1",
]
cluster = [
    "dask==2025.5.1",
    "ray==2.46.0",
]

[project.scripts]
cosmic-infant = "cosmic_mycelium.scripts.run_infant:main"
cosmic-cluster = "cosmic_mycelium.scripts.run_cluster:main"
cosmic-physics = "cosmic_mycelium.scripts.benchmark_physics:main"
cosmic-stress = "cosmic_mycelium.scripts.stress_test_physics:main"
cosmic-scale = "cosmic_mycelium.scripts.scale_test:main"

[tool.hatch.build.targets.wheel]
packages = ["cosmic_mycelium"]

# ============================================================================
# Tool configurations
# ============================================================================

[tool.black]
line-length = 88
target-version = ["py313"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | venv
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["**/.venv/*", "**/venv/*", "**/build/*"]

[tool.ruff]
target-version = "py313"
line-length = 88
exclude = [
    ".git",
    ".venv",
    "venv",
    "build",
    "dist",
    "__pycache__",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "PTH", # flake8-use-pathlib
    "PYI", # flake8-pyi
    "PT",  # flake8-pytest-style
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # function call in default args
    "C901",  # too complex (we use our own Cyclomatic check)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011", "PLR2004", "S101"]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_context = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "cosmic_mycelium.scripts.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = """
    -v
    --strict-markers
    --tb=short
    --durations=10
    --cov=cosmic_mycelium
    --cov-report=term-missing:skip-covered
    --cov-report=html:htmlcov
    --cov-report=xml:coverage.xml
    --cov-fail-under=80
"""
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "physics: physics validation tests",
    "slow: slow tests",
]

[tool.coverage.run]
branch = true
source = ["cosmic_mycelium"]
omit = [
    "tests/*",
    "cosmic_mycelium/scripts/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
]

# Cyclomatic complexity check (via mccabe, enforced via ruff)
[tool.ruff.lint.mccabe]
max-complexity = 10
